数据库实训/managment-system/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0

